修改檔案

+++ 貪食蛇.py
@@ -4,8 +4,7 @@ import sys
 import random
 import time
 from firebase import firebase
-
-
+import os
 
 A=np.zeros([50,50])
 for i in range(3,6):
@@ -286,20 +285,19 @@ win.blit(word1,(220,230))
 win.blit(word2,(220,270))
 win.blit(word3,(220,310))
 
-try:
+if not os.system('ping www.google.com'):
     db_url='https://game-3fb9c.firebaseio.com/'
     fdb=firebase.FirebaseApplication(db_url,None)
     h=fdb.get('/highest',None)
 
     for key in h:
         highest=h[key]
-
     if score>highest:
         fdb.delete('/highest',None)
         fdb.post('/highest',score)
         word4=font.render('new record',False,(255,0,0))
         win.blit(word4,(225,350))
-except:
+else:
     pass
 pygame.display.update()
